==> notebooks/04_inference_calculation.ipynb <==
import pandas as pd
import numpy as np
import joblib
import logging
import time

inference_data = pd.read_csv('../data/raw/test.csv')
inference_data.drop(['Id'], axis =1, inplace=True)
inference_data.head()

house_prices_pipeline = joblib.load('../models/house_prices_pipeline.pkl')
house_prices_pipeline

predicciones = house_prices_pipeline.predict(inference_data)
